Configurable comparison metric
The original plan is to use SSIM, but using SSIM to compare images takes a lot
more time than that of MSE (can be over 300%). The comparison metric is now
configurable, user can use SSIM if it does not matter to take more time to do
the job, or use MSE if it is more important to get the differences in shorter
period of time.