Continue to compare rest of the images even when a comparison failed
In some rare cases, the image cannot be compared successfully by ImageMagick or
FFmpeg due to corrupted file header. The previous implementation reports an
error and stop the comparisons, which forces the user to remove the problematic
image files and restart the comparisons, this can be time-consuming. The new
implementation reports an error, consider the pair to be abnormal, and continue
comparing the rest of the images.